file-filter: extract and simplify handling of project references
